Cast reconstructed streams to backbone dtype
Return FP32-codec reconstructions in the frozen backbone activation dtype so BF16 decoder cross-attention accepts receiver memory during direct evaluation. Add a real tiny T5Gemma regression and validate with 123 tests, Ruff and Pyright.

+++ jscc/models/split_model.py
@@ -139,7 +139,11 @@ class SplitModel(nn.Module):
             film_snr = self.snr_db
             if film_snr is None:
                 film_snr = self.channel_config["clean_film_snr"]
-            return codec.decode(received, film_snr)
+            # LayerNorm and residual blocks remain FP32 parameters in the
+            # corrected precision contract.  Return the reconstructed stream
+            # in the backbone activation dtype before downstream BF16 linear
+            # layers consume it.
+            return codec.decode(received, film_snr).to(dtype=hidden.dtype)
 
     def _roundtrip(self, hidden):
         self.activation = hidden.detach()
